# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ticket Sell API",
    "description": "API for selling tickets. All responses follow the envelope pattern: `{ success, data, message, code }`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Default server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "description": "Response payload. Shape varies by endpoint."
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable success code in SCREAMING_SNAKE_CASE",
            "example": "USER_CREATED"
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "code"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "data": {
            "type": "array",
            "items": {},
            "example": []
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials, verify the email or password"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code in SCREAMING_SNAKE_CASE",
            "example": "INVALID_CREDENTIALS"
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "code"
        ]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid Bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "data": [],
              "message": "Authentication required. Please provide a valid Bearer token.",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request body failed schema validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "data": [],
              "message": "One or more fields are invalid or missing.",
              "code": "VALIDATION_ERROR"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "description": "Creates a new user account. The password is stored hashed and is never returned.",
        "operationId": "authRegister",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "example": "S3cr3tP@ss"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "name": "John Doe",
                    "email": "john@example.com"
                  },
                  "message": "User created successfully.",
                  "code": "USER_CREATED"
                }
              }
            }
          },
          "409": {
            "description": "Email address is already associated with an existing account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "data": [],
                  "message": "This email address is already in use. Please try a different one.",
                  "code": "EMAIL_ALREADY_IN_USE"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Authenticate a user",
        "description": "Validates credentials and returns a short-lived JWT access token and a long-lived refresh token.",
        "operationId": "authLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "S3cr3tP@ss"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful. Tokens are returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refreshToken": "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."
                  },
                  "message": "Login successful.",
                  "code": "LOGIN_SUCCESS"
                }
              }
            }
          },
          "401": {
            "description": "Credentials are incorrect.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "data": [],
                  "message": "Invalid credentials, verify the email or password.",
                  "code": "INVALID_CREDENTIALS"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "description": "Exchanges a valid refresh token for a new short-lived JWT access token.",
        "operationId": "authRefresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4..."
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  },
                  "message": "Token refreshed successfully.",
                  "code": "TOKEN_REFRESHED"
                }
              }
            }
          },
          "401": {
            "description": "The provided refresh token is expired, revoked, or malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "data": [],
                  "message": "The refresh token is invalid or has expired. Please log in again.",
                  "code": "INVALID_REFRESH_TOKEN"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get authenticated user profile",
        "description": "Returns the profile of the currently authenticated user identified by the Bearer token.",
        "operationId": "usersGetMe",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "name": "John Doe",
                    "email": "john@example.com"
                  },
                  "message": "User profile fetched successfully.",
                  "code": "USER_FETCHED"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/me/tickets": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "List the authenticated user's tickets",
        "description": "Returns open reservations and completed purchases for the user identified by the Bearer token. `RESERVED` items come from reservations; `SOLD` items come from purchases.",
        "operationId": "usersGetMyTickets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Reservations and purchases returned successfully. `data` is an empty array when the user has neither.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "clx0reserv0001abcdefghijk",
                      "ticketId": "clx0ticket0001abcdefghijk",
                      "userId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "quantity": 2,
                      "status": "RESERVED",
                      "createdAt": "2026-09-24T12:30:00.000Z",
                      "ticket": {
                        "id": "clx0ticket0001abcdefghijk",
                        "title": "Rock Festival 2026",
                        "description": "Main stage, general admission",
                        "price": 150,
                        "quantity": 100,
                        "createdAt": "2026-09-24T12:00:00.000Z",
                        "updatedAt": "2026-09-24T12:00:00.000Z"
                      }
                    },
                    {
                      "id": "clx0purch00001abcdefghijk",
                      "ticketId": "clx0ticket0001abcdefghijk",
                      "userId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "quantity": 1,
                      "status": "SOLD",
                      "createdAt": "2026-09-24T12:45:00.000Z",
                      "ticket": {
                        "id": "clx0ticket0001abcdefghijk",
                        "title": "Rock Festival 2026",
                        "description": "Main stage, general admission",
                        "price": 150,
                        "quantity": 100,
                        "createdAt": "2026-09-24T12:00:00.000Z",
                        "updatedAt": "2026-09-24T12:00:00.000Z"
                      }
                    }
                  ],
                  "message": "User tickets fetched successfully.",
                  "code": "USER_TICKETS_FETCHED"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}